+++ common/utils_data_generation.hpp
@@ -159,7 +159,9 @@ struct generate_limits
 };
 
 template<typename T>
-struct generate_limits<T, std::enable_if_t<rocprim::is_integral<T>::value>>
+struct generate_limits<
+    T,
+    std::enable_if_t<rocprim::traits::get<T>::is_build_in() && rocprim::is_integral<T>::value>>
 {
     static inline T min()
     {
@@ -172,7 +174,9 @@ struct generate_limits<T, std::enable_if_t<rocprim::is_integral<T>::value>>
 };
 
 template<typename T>
-struct generate_limits<T, std::enable_if_t<rocprim::is_floating_point<T>::value>>
+struct generate_limits<T,
+                       std::enable_if_t<rocprim::traits::get<T>::is_build_in()
+                                        && rocprim::is_floating_point<T>::value>>
 {
     static inline T min()
     {